fix(node/engine): derivation sync start
