+++ crates/node/service/src/lib.rs
@@ -21,6 +21,3 @@ pub use actors::{
     InboundDerivationMessage, L1WatcherRpc, L1WatcherRpcError, NetworkActor, NetworkActorError,
     NodeActor, RpcActor, RpcActorError, RuntimeActor, RuntimeLauncher,
 };
-
-mod sync_start;
-pub use sync_start::{L2ForkchoiceState, SyncStartError, find_starting_forkchoice};
