+++ crates/node/sources/Cargo.toml
@@ -24,6 +24,7 @@ kona-protocol.workspace = true
 kona-providers-alloy.workspace = true
 
 # Alloy
+alloy-eips.workspace = true
 alloy-provider.workspace = true
 alloy-transport.workspace = true
 alloy-primitives.workspace = true
