+++ crates/node/sources/src/lib.rs
@@ -10,6 +10,9 @@
 #[macro_use]
 extern crate tracing;
 
+mod sync;
+pub use sync::{L2ForkchoiceState, SyncStartError, find_starting_forkchoice};
+
 mod runtime;
 pub use runtime::{RuntimeConfig, RuntimeLoader, RuntimeLoaderError};
